refactor: move spy core out of protocol/pubsub into the spy tool package
The spy core (TopicStats, SpySource, LCM/Zenoh sources, TransportSpy) is an
observability tool built on top of the pubsub layer, not part of it — it only
depends on the pubsub impls, and its only consumers are the spy CLI and tests.
Sitting in protocol/pubsub/ was a layering inversion. Move it next to run_spy:

  dimos/protocol/pubsub/spy.py       -> dimos/utils/cli/spy/core.py
  dimos/protocol/pubsub/test_spy.py  -> dimos/utils/cli/spy/test_spy.py

Imports updated; no behavior change. pubsub suite and spy tests green.
